requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
kombu==5.3.5
prompt-toolkit==3.0.43
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
six==1.16.0
slack-bolt==1.18.1
slack_sdk==3.27.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
